rebasing the code by seperating similar endpoints into their respective routes. created the user routes at this point
