+++ backend/app/emails.py
@@ -47,7 +47,7 @@ async def send_email(email: List, instance: User):
                     <p>Thank you for registering with us. Kindly click on the link below to
                     verify your email and have full acccess to the platform.</p>
 
-                    <a href="https://heed.hng.tech/emailverification?token={token}">Verify your email address </a>
+                    <a href="https://api.heed.hng.tech/verification?token={token}">Verify your email address </a>
         </div>
     """
