+++ backend/app/routers/users.py
@@ -1,0 +1,260 @@
+from fastapi import FastAPI, status, Depends, APIRouter,  UploadFile, File, Form, Query, Request
+from typing import List, Union, Optional
+import services as _services
+import models, schema
+from sqlalchemy.orm import Session
+import auth
+from . import utility as utils
+from fastapi.responses import JSONResponse
+from fastapi.encoders import jsonable_encoder
+import os
+import cloudinary
+import cloudinary.uploader
+from BitlyAPI import shorten_urls
+import crud
+from jwt import main_login, get_access_token, verify_password, refresh
+from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from emails import send_email, verify_token, send_password_reset_email, password_verif_token
+
+
+
+
+user_router = APIRouter(
+    prefix='/users',
+    tags=['users'],
+)
+
+
+# endpoint for user login
+@user_router.post('/login', summary = "create access token for logged in user",
+                  status_code= status.HTTP_200_OK)
+async def login(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(_services.get_session)):
+    # return token once the user has been successfully authenticated, or it returns an error.
+    return await jwt.main_login(form_data, db)
+
+
+# creating a users account.
+@user_router.post("/create_users", status_code= status.HTTP_201_CREATED, 
+                  summary = "create/register a new user user", response_model=schema.User)
+async def create_user(user: schema.UserCreate, db: Session = Depends(_services.get_session)):
+    db_user = crud.get_user_by_email(db, email=user.email)
+    
+    # if user exists, throw an exception.
+    if db_user:
+        raise HTTPException(status_code=400, detail="Email already registered")
+    try:
+        # create the user before sending a mail.
+        new_user = crud.create_user(db=db, user=user)
+        await send_email([user.email], user)
+    except Exception as e:
+        return JSONResponse(
+            status_code= status.HTTP_400_BAD_REQUEST,
+            content=jsonable_encoder({"detail": str(e)}),
+        )
+        
+    return {    
+        "detail" : new_user
+    }
+
+# get all users, only available for the admin end.
+@user_router.get("/get_all_users", summary = "get all users", status_code= 200, response_model=list[schema.User])
+def read_users(skip: int = 0, limit: int = 100, db: Session = Depends(_services.get_session), 
+               user: models.User = Depends(auth.get_admin)):
+    try:
+        users = crud.get_users(db, skip=skip, limit=limit)
+    except Exception as e:
+       return JSONResponse(
+            status_code= status.HTTP_400_BAD_REQUEST,
+            content=jsonable_encoder({"detail": str(e)}),
+        )
+    return {      
+        detail: users
+    }
+    
+@user_router.get("/get_user/{user_id}", summary = "get user by id", status_code=200, response_model=schema.User)
+def read_user(user_id: int, db: Session = Depends(_services.get_session)):
+    try:
+        db_user = crud.get_user(db, user_id=user_id)
+        if db_user is None:
+            raise HTTPException(status_code=404, detail="User not found")
+    except Exception as e:
+        return JSONResponse(
+            status_code= status.HTTP_400_BAD_REQUEST,
+            content=jsonable_encoder({"detail": str(e)}),
+        )
+    
+    return {
+        "detail": db_user
+    }
+    
+@user_router.patch("/update_profile", summary="Update user profile details", status_code=status.HTTP_200_OK)
+def update_profile( 
+                   firstname:Optional[str] = Form(None), 
+                   lastname:Optional[str] = Form(None), 
+                   company_name: Optional[str] = Form(None), 
+                   company_address:Optional[str] = Form(None), 
+                   phone_number: Optional[str] = Form(None), 
+                   db:Session = Depends(_services.get_session), 
+                   current_user:schema.User = Depends(auth.get_active_user),
+                  image_file: Optional[UploadFile] = File(None, description="Company Profile Image/Logo")):
+    user_profile = db.query(models.UserProfile).filter(models.UserProfile.id == current_user.id).first()
+    if user_profile is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                        detail=f"The Profile for user with id {current_user.id} does not exist")
+        
+    if user_profile.id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN , 
+                                    detail="Not authorized to perform requested action")
+        
+    user = db.query(models.User).filter(models.User.id == user_profile.id).first()
+    company = db.query(models.Company).filter(models.Company.id == user.company_id).first()
+    field_names = [company_name, company_address, phone_number, firstname, lastname, image_file]
+    for field in field_names:
+        if field == None:
+            pass
+        elif field is company_name:
+            company.name = company_name
+        elif field is company_address:
+            user_profile.company_address = company_address
+        elif field is phone_number:
+            user_profile.phone_number = phone_number
+        elif field is firstname:
+            user.first_name = firstname  
+        elif field is lastname:
+           user.last_name = lastname
+        else:
+            try:
+                image_response = cloudinary.uploader.upload(image_file.file)
+                image_url = image_response.get("secure_url") 
+                user_profile.company_logo_url = image_url        
+            except Exception:
+                raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="There was an error uploading the file")
+
+    db.commit() 
+    db.refresh(company)
+    db.refresh(user_profile)
+    return {
+        "detail": user_profile
+    }
+    
+
+@user_router.get('/verification', summary = "verify a user by email", status_code = status.HTTP_200_OK)
+async def email_verification(request: Request, token: str, db: Session = Depends(_services.get_session)):
+
+    user = await verify_token(token, db)
+
+
+    if user and not user.is_active:
+        user.is_active = True
+        user.is_verified = True
+        db.commit()
+        return{
+            "detail" : f"Hello {user.first_name}, your account has been successfully verified"}
+    return {
+        "detail" : f"Hello {user.first_name}, you already have an active account with Heed!"
+    }
+    
+    
+@user_router.get("/account", summary = "get user profile details", tags=['users'])
+async def my_profile (db: Session = Depends(_services.get_session), user: models.User = Depends(auth.get_active_user)):
+    user_id = user.id
+    try:
+        return crud.get_user_profile(db, user_id)
+    except Exception as e:
+        return JSONResponse(
+            status_code= status.HTTP_400_BAD_REQUEST,
+            content=jsonable_encoder({"detail": str(e)}),
+        )
+
+
+
+@user_router.post('/forgot-password', summary = "get token for password reset", status_code = 200)
+async def forgot_password(email: schema.ForgetPassword, db: Session = Depends(_services.get_session)):
+    
+    try:
+        user: models.User = crud.get_user_by_email(db, email.email)
+
+        if user is None:
+            raise HTTPException(status_code=404, detail="User not found")
+
+        
+        token = await send_password_reset_email([email.email], user)
+    except Exception as e:
+        return JSONResponse(
+            status_code= status.HTTP_400_BAD_REQUEST,
+            content=jsonable_encoder({"detail": str(e)}),
+        )
+    return {"detail": token}
+    
+
+@user_router.patch('/reset-password', summary = "reset password", status_code = 200)
+async def reset_password(token: str, new_password: schema.UpdatePassword, db: Session = Depends(_services.get_session)):
+    
+    try:
+        email = password_verif_token(token)
+        user: models.User = crud.get_user_by_email(db, email)
+            
+        if user is None:
+            raise HTTPException(status_code=404, detail="User not found")
+
+        its_match = verify_password(new_password.password, user.password)
+        its_le_eight = len(new_password.password) < 8
+
+        if its_match:
+            raise HTTPException(status_code=500, detail="New password cannot be the same as old password")
+        elif its_le_eight:
+            raise HTTPException(status_code=500, detail="Password must have at least 8 characters")
+
+        
+        reset_done = crud.reset_password(db, new_password.password, user)
+
+        if reset_done is None:
+            raise HTTPException(status_code=500, detail="Failed to update password")
+        
+    except Exception as e:
+        return JSONResponse(
+            status_code= status.HTTP_400_BAD_REQUEST,
+            content=jsonable_encoder({"detail": str(e)}),
+        )
+    
+    return {
+        "detail":reset_done
+        }
+
+
+
+@user_router.patch('/change-password', summary = "change password", status_code = 200)
+async def change_password(password_schema: schema.ChangePassword, 
+                          db: Session = Depends(_services.get_session), user: models.User = Depends(auth.get_active_user)):
+    try:
+        its_match = password_schema.old_password == password_schema.new_password
+        its_le_eight = len(password_schema.new_password) < 8
+
+        if its_match:
+            raise HTTPException(status_code=500, detail="New password cannot be the same as old password")
+        elif its_le_eight:
+            raise HTTPException(status_code=500, detail="Password must have at least 8 characters")
+
+        user_db = crud.get_user_by_email(db, user.email)
+
+        if user_db is None:
+            raise HTTPException(status_code=404, detail="User not found")
+
+        password_match =  verify_password(password_schema.old_password, user_db.password)
+
+        if not password_match:
+            raise HTTPException(status_code=400, detail="Password does not match")
+        
+        reset_done = crud.reset_password(db, password_schema.new_password, user_db)
+
+        if reset_done is None:
+            raise HTTPException(status_code=400, detail="Failed to update password")
+    except Exception as e:
+        return JSONResponse(
+            status_code= status.HTTP_400_BAD_REQUEST,
+            content=jsonable_encoder({"detail": str(e)}),
+        )
+    
+    return {
+        "detail": reset_done
+    }
